OLD_FILES: don't try to delete /etc/ssh
In our systems, /etc/ssh should always be a symbolic link to
/usr/local/etc/ssh (because that's easier than teaching Puppet
to look in the right place). There's no way for bmake to test
whether that is the case (at the time the list of "old files"
is being assembled, we don't even know what ${DESTDIR} is)
so just err on the side of caution and don't try to delete them
even if /etc/ssh should happen to be a real directory and not
a symlink like it's supposed to be.