kernel configuration: remove outdated syscons configuration
We have not built the CSAIL-SC kernel in a long time and were not
keeping it up to date with other kernel changes.  Remove the config
file.  Remove comments in the CSAIL kernel implying that it was a
temporary file (from the days when it was CSAIL-VT and CSAIL-SC was
still CSAIL).